room_classifier_api/requirements.txt
fastapi==0.124.2
uvicorn==0.38.0
tensorflow==2.18.0
pillow==12.0.0
numpy==1.26.4
python-multipart==0.0.20